requirements.txt
# Core dependencies
streamlit==1.31.0
anthropic==0.18.1
python-dotenv==1.0.1

# Document processing
python-docx==1.1.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data handling
pandas==2.2.0
openpyxl==3.1.2

# Additional utilities
Pillow==10.2.0
